+++ extensions/azure/blobstorage/blob-core/src/testFixtures/java/org/eclipse/dataspaceconnector/azure/blob/core/AzureStorageTestFixtures.java
@@ -20,34 +20,33 @@ import org.eclipse.dataspaceconnector.spi.types.domain.transfer.DataFlowRequest;
 
 public class AzureStorageTestFixtures {
 
-    private static final Faker FAKER = new Faker();
+    static Faker faker = new Faker();
 
     public static DataFlowRequest.Builder createRequest(String type) {
         return DataFlowRequest.Builder.newInstance()
-                .id(FAKER.internet().uuid())
-                .processId(FAKER.internet().uuid())
-                .sourceDataAddress(createDataAddress(type).build())
-                .destinationDataAddress(createDataAddress(type).build());
-    }
-
-    public static DataAddress.Builder createDataAddress(String type) {
-        return DataAddress.Builder.newInstance().type(type);
+                .id("1").processId("1")
+                .sourceDataAddress(DataAddress.Builder.newInstance().type(type).build())
+                .destinationDataAddress(DataAddress.Builder.newInstance().type(type).build());
     }
 
     public static String createAccountName() {
-        return FAKER.lorem().characters(3, 24, false, false);
+        return faker.lorem().characters(3, 24, false, false);
     }
 
     public static String createContainerName() {
-        return FAKER.lorem().characters(3, 40, false, false);
+        return faker.lorem().characters(3, 40, false, false);
     }
 
     public static String createBlobName() {
-        return FAKER.lorem().characters(3, 40, false, false);
+        return faker.lorem().characters(3, 40, false, false);
     }
 
     public static String createSharedKey() {
-        return FAKER.lorem().characters();
+        return faker.lorem().characters();
+    }
+
+    public static String createSharedAccessSignature() {
+        return faker.lorem().characters();
     }
 
     private AzureStorageTestFixtures() {