+++ extensions/azure/data-plane/storage/src/test/java/org/eclipse/dataspaceconnector/azure/dataplane/azurestorage/pipeline/AzureStorageDataSinkTest.java
@@ -27,7 +27,9 @@ import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.util.List;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
@@ -39,7 +41,10 @@ import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFix
 import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createBlobName;
 import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createContainerName;
 import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createRequest;
-import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createSharedKey;
+import static org.eclipse.dataspaceconnector.azure.blob.core.AzureStorageTestFixtures.createSharedAccessSignature;
+import static org.eclipse.dataspaceconnector.azure.dataplane.azurestorage.pipeline.TestFunctions.sharedAccessSignatureMatcher;
+import static org.mockito.ArgumentMatchers.argThat;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -54,7 +59,8 @@ class AzureStorageDataSinkTest {
 
     String accountName = createAccountName();
     String containerName = createContainerName();
-    String sharedKey = createSharedKey();
+    String sharedAccessSignature = createSharedAccessSignature();
+
     String blobName = createBlobName();
     String content = faker.lorem().sentence();
 
@@ -63,25 +69,35 @@ class AzureStorageDataSinkTest {
     AzureStorageDataSink dataSink = AzureStorageDataSink.Builder.newInstance()
             .accountName(accountName)
             .containerName(containerName)
-            .sharedKey(sharedKey)
+            .sharedAccessSignature(sharedAccessSignature)
             .requestId(request.build().getId())
             .blobStoreApi(blobStoreApi)
             .executorService(executor)
             .monitor(monitor)
             .build();
     BlobAdapter destination = mock(BlobAdapter.class);
+    BlobAdapter completionMarker = mock(BlobAdapter.class);
     InputStreamDataSource part = new InputStreamDataSource(blobName, new ByteArrayInputStream(content.getBytes(UTF_8)));
     ByteArrayOutputStream output = new ByteArrayOutputStream();
+    OutputStream completionMarkerOutput = mock(OutputStream.class);
 
     @BeforeEach
     void setUp() {
         when(destination.getOutputStream()).thenReturn(output);
         when(blobStoreApi.getBlobAdapter(
-                accountName,
-                containerName,
-                blobName,
-                sharedKey))
+                eq(accountName),
+                eq(containerName),
+                eq(blobName),
+                sharedAccessSignatureMatcher(sharedAccessSignature)))
                 .thenReturn(destination);
+
+        when(completionMarker.getOutputStream()).thenReturn(completionMarkerOutput);
+        when(blobStoreApi.getBlobAdapter(
+                eq(accountName),
+                eq(containerName),
+                argThat(s -> s.endsWith(".complete")),
+                sharedAccessSignatureMatcher(sharedAccessSignature)))
+                .thenReturn(completionMarker);
     }
 
     @Test
@@ -94,10 +110,10 @@ class AzureStorageDataSinkTest {
     @Test
     void transferParts_whenBlobClientCreationFails_fails() {
         when(blobStoreApi.getBlobAdapter(
-                accountName,
-                containerName,
-                blobName,
-                sharedKey))
+                eq(accountName),
+                eq(containerName),
+                eq(blobName),
+                sharedAccessSignatureMatcher(sharedAccessSignature)))
                 .thenThrow(exception);
         assertThatTransferPartsFails(part, "Error creating blob for %s on account %s", blobName, accountName);
     }
@@ -106,7 +122,6 @@ class AzureStorageDataSinkTest {
     void transferParts_whenWriteFails_fails() {
         when(destination.getOutputStream()).thenThrow(exception);
     }
-
 
     @Test
     void transferParts_whenReadFails_fails() {
@@ -127,6 +142,17 @@ class AzureStorageDataSinkTest {
         assertThatTransferPartsFails(part, "Error transferring blob for %s on account %s", blobName, accountName);
     }
 
+    @Test
+    void complete() throws IOException {
+        dataSink.complete();
+        verify(blobStoreApi).getBlobAdapter(
+                eq(accountName),
+                eq(containerName),
+                argThat(s -> s.endsWith(".complete")),
+                sharedAccessSignatureMatcher(sharedAccessSignature));
+        verify(completionMarkerOutput).close();
+    }
+
     private void assertThatTransferPartsFails(Part part, String logMessage, Object... args) {
         String message = format(logMessage, args);
         var result = dataSink.transferParts(List.of(part));
@@ -134,4 +160,5 @@ class AzureStorageDataSinkTest {
         assertThat(result.getFailureMessages()).containsExactly(message);
         verify(monitor).severe(message, exception);
     }
+
 }