+++ extensions/data-plane/data-plane-framework/src/main/java/org/eclipse/dataspaceconnector/dataplane/framework/DataPlaneFrameworkExtension.java
@@ -33,7 +33,6 @@ import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.Provides;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 import java.util.concurrent.Executors;
@@ -41,7 +40,7 @@ import java.util.concurrent.Executors;
 /**
  * Provides core services for the Data Plane Framework.
  */
-@Provides({DataPlaneManager.class, PipelineService.class, DataTransferExecutorServiceContainer.class, TransferServiceRegistry.class})
+@Provides({DataPlaneManager.class, PipelineService.class, DataTransferExecutorServiceContainer.class})
 public class DataPlaneFrameworkExtension implements ServiceExtension {
     private static final int IN_MEMORY_STORE_CAPACITY = 1000;
 
@@ -61,13 +60,13 @@ public class DataPlaneFrameworkExtension implements ServiceExtension {
     private static final String TRANSFER_THREADS = "edc.dataplane.transfer.threads";
     private static final int DEFAULT_TRANSFER_THREADS = 10;
 
+    private ServiceExtensionContext context;
+
     private DataPlaneManagerImpl dataPlaneManager;
+    private Monitor monitor;
 
     @Inject(required = false)
     private TransferServiceSelectionStrategy transferServiceSelectionStrategy;
-
-    @Inject(required = false)
-    private DataPlaneStore store;
 
     @Inject
     private ExecutorInstrumentation executorInstrumentation;
@@ -79,6 +78,7 @@ public class DataPlaneFrameworkExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
+        this.context = context;
         var pipelineService = new PipelineServiceImpl();
         pipelineService.registerFactory(new OutputStreamDataSinkFactory()); // Added by default to support synchronous data transfer, i.e. pull data
         context.registerService(PipelineService.class, pipelineService);
@@ -95,7 +95,9 @@ public class DataPlaneFrameworkExtension implements ServiceExtension {
                 executorInstrumentation.instrument(executorService, "Data plane transfers"));
         context.registerService(DataTransferExecutorServiceContainer.class, executorContainer);
 
-        Monitor monitor = context.getMonitor();
+        monitor = context.getMonitor();
+        var telemetry = context.getTelemetry();
+
         var queueCapacity = context.getSetting(QUEUE_CAPACITY, DEFAULT_QUEUE_CAPACITY);
         var workers = context.getSetting(WORKERS, DEFAULT_WORKERS);
         var waitTimeout = context.getSetting(WAIT_TIMEOUT, DEFAULT_WAIT_TIMEOUT);
@@ -107,24 +109,20 @@ public class DataPlaneFrameworkExtension implements ServiceExtension {
                 .waitTimeout(waitTimeout)
                 .pipelineService(pipelineService)
                 .transferServiceRegistry(transferServiceRegistry)
-                .store(registerStore(context))
-                .monitor(monitor).build();
+                .store(new InMemoryDataPlaneStore(IN_MEMORY_STORE_CAPACITY))
+                .monitor(monitor)
+                .telemetry(telemetry)
+                .build();
 
         context.registerService(DataPlaneManager.class, dataPlaneManager);
     }
 
-    @NotNull
-    private DataPlaneStore registerStore(ServiceExtensionContext context) {
-        if (store != null) {
-            return store;
-        }
-        var inMemoryStore = new InMemoryDataPlaneStore(IN_MEMORY_STORE_CAPACITY);
-        context.registerService(DataPlaneStore.class, inMemoryStore);
-        return inMemoryStore;
-    }
-
     @Override
     public void start() {
+        if (!context.hasService(DataPlaneStore.class)) {
+            monitor.info("Registering in-memory Data Plane store.");
+            context.registerService(DataPlaneStore.class, new InMemoryDataPlaneStore(IN_MEMORY_STORE_CAPACITY));
+        }
         dataPlaneManager.start();
     }